+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.25.0...HEAD
 
+### Added
+
+- Probes to work with the incident center
+
 ## [0.25.0][] - 2023-10-27
 
 [0.25.0]: https://github.com/chaostoolkit-incubator/chaostoolkit-aws/compare/0.24.0...0.25.0